Corrigido nome do pacote e mapeamento e adicionada funcionalidade de pegar data do ultimo login
