+++ WebContent/createPost.jsp
@@ -13,7 +13,7 @@
         <pre>
             Titulo : <input type="text" name="titulo">
                      <input type="text" name="texto">
-                     <select multiple size="3" name="tags">
+                     <select multiple class="select-checkbox" name="tags">
                      	<option value="EDUCACAO">Educação</option>
                      	<option value="SAUDE">Saúde</option>
                      	<option value="ARTIGO">Artigo</option>
