+++ WebContent/userPost.jsp
@@ -20,7 +20,7 @@
 		<c:if test="${posts.isEmpty() == false}">
 			<c:forEach items="${posts}" var="post">
 				<form action="<c:url value='ShowPost'/>" method="get">
-					<h3>Titulo: ${post.getTitle()} </h3>
+					<h3>Titulo: ${post.getTitle()}</h3>
 					<label><b>Tema:</b> ${post.getTags()}</label><br><br>
 					<label><b>Conteúdo:</b></label>
 					<div class="conteudo-postagem">
@@ -28,17 +28,6 @@
 					</div>						
 					<button type="submit" name ="id" value ="${post.getPostId()}" class="btn-verPost">Ver postagem completa</button><br><br> 	
 				</form>	
-			
-				<!--  <div>
-					<hr>
-					<h3>
-						quando clica se coloca o objeto post como atributo da sessao http, ou era pra estar acontecendo isso
-						<a href="post.jsp?id=${post.getPostId()}">
-							<c:out value="${post.getTitle()}" />
-						</a>
-					</h3>
-					<p><c:out value="${post.getText()}" /></p>
-				</div> -->
 			</c:forEach>
 		</c:if>
 	</div>
