+++ src/com/school/koren/controller/CommentPost.java
@@ -12,12 +12,12 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
-import com.school.koren.dao.AccountHome;
-import com.school.koren.dao.CommentaryHome;
-import com.school.koren.dao.PostHome;
 import com.school.koren.model.Account;
 import com.school.koren.model.Commentary;
 import com.school.koren.model.Post;
+import com.school.koren.repository.AccountHome;
+import com.school.koren.repository.CommentaryHome;
+import com.school.koren.repository.PostHome;
 
 /**
  * Servlet implementation class CommentPost
