+++ src/com/school/koren/controller/CreatePost.java
@@ -12,12 +12,12 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
-import com.school.koren.dao.AccountHome;
-import com.school.koren.dao.CategoryHome;
-import com.school.koren.dao.PostHome;
 import com.school.koren.model.Account;
 import com.school.koren.model.Post;
 import com.school.koren.model.Post.Tag;
+import com.school.koren.repository.AccountHome;
+import com.school.koren.repository.CategoryHome;
+import com.school.koren.repository.PostHome;
 
 /**
  * Servlet implementation class Post
