+++ src/com/school/koren/controller/DeletePost.java
@@ -11,8 +11,8 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
-import com.school.koren.dao.PostHome;
 import com.school.koren.model.Post;
+import com.school.koren.repository.PostHome;
 
 /**
  * Servlet implementation class DeletePost
