+++ src/com/school/koren/controller/ShowPost.java
@@ -11,10 +11,10 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
-import com.school.koren.dao.CommentaryHome;
-import com.school.koren.dao.PostHome;
 import com.school.koren.model.Commentary;
 import com.school.koren.model.Post;
+import com.school.koren.repository.CommentaryHome;
+import com.school.koren.repository.PostHome;
 
 /**
  * Servlet implementation class ShowPost
