+++ src/com/school/koren/controller/ZeroADois.java
@@ -12,10 +12,10 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
-import com.school.koren.dao.AccountHome;
-import com.school.koren.dao.CategoryHome;
-import com.school.koren.dao.PostHome;
 import com.school.koren.model.Post;
+import com.school.koren.repository.AccountHome;
+import com.school.koren.repository.CategoryHome;
+import com.school.koren.repository.PostHome;
 
 /**
  * Servlet implementation class SeteADez
