+++ src/com/school/koren/repository/CategoryHome.java
@@ -1,4 +1,4 @@
-package com.school.koren.dao;
+package com.school.koren.repository;
 
 import java.util.List;
 
@@ -15,12 +15,12 @@ import org.hibernate.SessionFactory;
 import org.hibernate.cfg.Configuration;
 import org.hibernate.query.Query;
 
-import com.school.koren.model.Account;
+import com.school.koren.model.Category;
 
 
-public class AccountHome {
+public class CategoryHome {
 
-	private static final Log log = LogFactory.getLog(AccountHome.class);
+	private static final Log log = LogFactory.getLog(CategoryHome.class);
 
 	private final SessionFactory sessionFactory = getSessionFactory();
 
@@ -34,9 +34,8 @@ public class AccountHome {
 		}
 	}
 
-
-	public void persist(Account transientInstance) {
-		log.debug("persisting Account instance");
+	public void persist(Category transientInstance) {
+		log.debug("persisting Category instance");
 		try {
 			Session session = sessionFactory.getCurrentSession();
 			session.beginTransaction();
@@ -50,8 +49,8 @@ public class AccountHome {
 		}
 	}
 
-	public void attachDirty(Account instance) {
-		log.debug("attaching dirty Account instance");
+	public void attachDirty(Category instance) {
+		log.debug("attaching dirty Category instance");
 		try {
 			sessionFactory.getCurrentSession().saveOrUpdate(instance);
 			log.debug("attach successful");
@@ -61,8 +60,8 @@ public class AccountHome {
 		}
 	}
 
-	public void attachClean(Account instance) {
-		log.debug("attaching clean Account instance");
+	public void attachClean(Category instance) {
+		log.debug("attaching clean Category instance");
 		try {
 			sessionFactory.getCurrentSession().lock(instance, LockMode.NONE);
 			log.debug("attach successful");
@@ -72,8 +71,8 @@ public class AccountHome {
 		}
 	}
 
-	public void delete(Account persistentInstance) {
-		log.debug("deleting Account instance");
+	public void delete(Category persistentInstance) {
+		log.debug("deleting Category instance");
 		try {
 			Session session = sessionFactory.getCurrentSession();
 			session.beginTransaction();
@@ -87,12 +86,12 @@ public class AccountHome {
 		}
 	}
 
-	public Account merge(Account detachedInstance) {
-		log.debug("merging Account instance");
+	public Category merge(Category detachedInstance) {
+		log.debug("merging Category instance");
 		try {
 			Session session = sessionFactory.getCurrentSession();
 			session.beginTransaction();
-			Account result = (Account) session.merge(detachedInstance);
+			Category result = (Category) session.merge(detachedInstance);
 			session.getTransaction().commit();
 			session.close();
 			log.debug("merge successful");
@@ -103,13 +102,13 @@ public class AccountHome {
 		}
 	}
 
-	public Account findById(int id) {
-		log.debug("getting Account instance with id: " + id);
+	public Category findById(int id) {
+		log.debug("getting Category instance with id: " + id);
 		try {
 			Session session = sessionFactory.getCurrentSession();
 			session.beginTransaction();
 			
-			Account instance = (Account) session.get(Account.class, id);
+			Category instance = (Category) session.get(Category.class, id);
 			if (instance == null) {
 				log.debug("get successful, no instance found");
 			} else {
@@ -123,37 +122,25 @@ public class AccountHome {
 		}
 	}
 
-	public List<Account> findByExample(Account instance) {
-		log.debug("finding Account instance by example");
+	public List<Category> findByExample(Category instance) {
+		log.debug("finding Category instance by example");
 		try {
 			Session session = sessionFactory.getCurrentSession();
 			session.beginTransaction();
 			CriteriaBuilder builder = session.getCriteriaBuilder();
-			CriteriaQuery<Account> criteria = builder.createQuery(Account.class);
-			Root<Account> root = criteria.from(Account.class);
+			CriteriaQuery<Category> criteria = builder.createQuery(Category.class);
+			Root<Category> root = criteria.from(Category.class);
 			
 			Predicate predicate = builder.and();
 			
 			//Replicar para todos os atributos
-			if(instance.getUsername() != null) {
-				predicate = builder.and(builder.equal(root.get("username"), instance.getUsername()));
-			}
-			
-			if(instance.getEmail() != null) {
-				predicate = builder.and(builder.equal(root.get("email"), instance.getEmail()));
-			}
-			
-			if(instance.getCreationDate() != null) {
-				predicate = builder.and(builder.equal(root.get("creation_date"), instance.getCreationDate()));
-			}
-			
-			if(instance.getLastLogin() != null) {
-				predicate = builder.and(builder.equal(root.get("last_login"), instance.getLastLogin()));
+			if(instance.getTitle() != null) {
+				predicate = builder.and(builder.equal(root.get("title"), instance.getTitle()));
 			}
 			
 			criteria.select(root).where(predicate);
-			Query<Account> q = session.createQuery(criteria);
-			List<Account> results = q.getResultList();
+			Query<Category> q = session.createQuery(criteria);
+			List<Category> results = q.getResultList();
 			log.debug("find by example successful, result size: " + results.size());
 			session.close();
 			return results;
