+++ src/hibernate.cfg.xml
@@ -10,9 +10,9 @@
   <property name="hibernate.default_schema">koren</property>
   <property name="hibernate.dialect">org.hibernate.dialect.PostgreSQLDialect</property>
   <property name="hibernate.current_session_context_class">thread</property>
-  <mapping class="com.school.Account" file="Account" package="com.school"/>
-  <mapping class="com.school.Category" file="Category" package="com.school"/>
-  <mapping class="com.school.Commentary" file="Commentary" package="com.school"/>
-  <mapping class="com.school.Post" file="Post" package="com.school"/>
+  <mapping class="com.school.koren.model.Account" file="Account" package="com.school.koren.model"/>
+  <mapping class="com.school.koren.model.Category" file="Category" package="com.school.koren.model"/>
+  <mapping class="com.school.koren.model.Commentary" file="Commentary" package="com.school.koren.model"/>
+  <mapping class="com.school.koren.model.Post" file="Post" package="com.school.koren.model"/>
  </session-factory>
 </hibernate-configuration>
